:sparkles: feat: makeLineFromPoints method
Finnaly finished all the route maker function logic to match all possible lengths of the points array. Only make lines from a list with 2 or more Points and have optional logic for the fullName filter.

Changed the structure of the function that creates the Feature of the Points geometry to a proper function and given it a name: makeGeometryPointFromArrayList.